+++ entity/src/main/resources/db/migration/tables/validation/V1.81__add_rejection_reason_column.sql
@@ -1,0 +1,7 @@
+-- ------------------------------------------------
+-- Version: v1.81
+--
+-- Description: Migration that creates the validation rejection reason column in the Validation table
+-- -------------------------------------------------
+ALTER TABLE Validation
+ADD COLUMN rejection_reason VARCHAR(255);